Addressing some PROTOTYPE markers
